// types/anime.ts
export interface GetAnimeApiResponse {
  data: Anime[];
  meta: Meta3;
  links: Links18;
}

export interface Anime {
  id: string;
  type: string;
  links: Links;
  attributes: Attributes;
  relationships: Relationships;
}

export interface Links {
  self: string;
}

export interface Attributes {
  createdAt: string;
  updatedAt: string;
  slug: string;
  synopsis: string;
  description: string;
  coverImageTopOffset: number;
  titles: Titles;
  canonicalTitle: string;
  abbreviatedTitles: string[];
  averageRating: string;
  ratingFrequencies: RatingFrequencies;
  userCount: number;
  favoritesCount: number;
  startDate: string;
  endDate?: string;
  nextRelease: any;
  popularityRank: number;
  ratingRank: number;
  ageRating?: string;
  ageRatingGuide?: string;
  subtype: string;
  status: string;
  tba: any;
  posterImage: PosterImage;
  coverImage: CoverImage;
  episodeCount: number;
  episodeLength?: number;
  totalLength: number;
  youtubeVideoId?: string;
  showType: string;
  nsfw: boolean;
}
// ...
export interface Relationships {
  genres: Genres;
  categories: Categories;
  castings: Castings;
  installments: Installments;
  mappings: Mappings;
  reviews: Reviews;
  mediaRelationships: MediaRelationships;
  characters: Characters;
  staff: Staff;
  productions: Productions;
  quotes: Quotes;
  episodes: Episodes;
  streamingLinks: StreamingLinks;
  animeProductions: AnimeProductions;
  animeCharacters: AnimeCharacters;
  animeStaff: AnimeStaff;
}
// ...
export interface Meta3 {
  count: number;
}

export interface Links18 {
  first: string;
  next: string;
  last: string;
}